# Changes `selector-no-type` to automatically resolve nested selectors when `ignore: ["descendant"]` option is enabled.