--- a/src/rules/selector-no-type/README.md
+++ b/src/rules/selector-no-type/README.md
@@ -50,6 +50,10 @@ The following patterns are *not* considered warnings:
 
 ## Optional options
 
+### `resolveNestedSelectors: true | false` (default: `false`)
+
+This option will resolve nested selectors before checking any rules.
+
 ### `ignore: ["compounded", "descendant"]`
 
 #### `"compounded"`