### Description

This PR addresses #1654.

These changes will automatically resolve nested selectors automatically if the `ignore: ["descendant"]` option is enabled. 

And since this is my first PR on this project, please let me know if there's anything else I may have missed or of there is any other pertinent information I should add to the PR.

### Benchmarks

#### Before

```
❯ npm run benchmark-rule -- selector-no-type "[true, {\"ignore\": [\"descendant\"]}]"

Warnings: 406
Mean: 171.1881831515151 ms
Deviation: 38.763392889220015 ms
```

#### After

```
❯ npm run benchmark-rule -- selector-no-type "[true, {\"ignore\": [\"descendant\"]}]"

Warnings: 406
Mean: 156.6367916 ms
Deviation: 25.370745159382917 ms
```