@ntwb 2016-07-15
@manovotny Thanks, we'll need a changelog entry similar to [this commit](https://github.com/stylelint/stylelint/commit/c7bb78cebb3d2a14ef52669bb36104951b20c4c2#diff-4ac32a78649ca5bdd8e0ba38b7006a1e)

http://stylelint.io/developer-guide/rules/ also includes a few checklists when creating/updating rules