@manovotny 2016-07-15
@ntwb : : 

Fix added to changelog.

I reviewed the [checklist](http://stylelint.io/developer-guide/rules/#checklist) and I *think* I've got everything covered. Please let me know if you think I've missed something.