@ntwb 2016-07-15
@manovotny Can you add some SCSS syntax tests, similar [to these](https://github.com/stylelint/stylelint/blob/master/src/rules/selector-class-pattern/__tests__/index.js#L118-L133) please