@jeddy3 2016-07-15
@manovotny Thanks for contributing! Looking good so far.

I think we'll need to tweak the implementation a little though. In https://github.com/stylelint/stylelint/issues/1654#issuecomment-232754442 I said:

> I don't think we need to add a 2nd option this time as automatically resolving the nesting for an ignore secondary option won't be a breaking change.

And then @davidtheclark [categorised](https://github.com/stylelint/stylelint/issues/1654#issuecomment-232837182) as a bug accordingly.

As such, I think we can do away with the `resolveNestedSelectors` *option*. Instead, resolving the nesting is triggered automatically whenever the `ignore: []` secondary option is used. This change in default behaviour won't be a breaking change as we're making the rule more permissive.