@manovotny 2016-07-15
@jeddy3 : : 

Oh... I'm a dope. I totally misunderstood what you said. When you said `I don't think we need to add a 2nd option`, I thought you meant we didn't need to add something alongside `compounded` and  `descendant`, which I guess would be 3rd option. 

Anyway. My bad. 

We're on the same page now. I'll get those changes and the `syntax: "scss"` changes @ntwb wanted in the tests.

Thanks!