@manovotny 2016-07-15
Ok, here's were we're at...

I am now resolving nested selectors automatically if the `ignore: ["descendant"]` option is enabled. I am only doing this for `descendant` because `compounded` will work as expected regardless if it's nested or not. 

And because nesting is only possible with a preprocessor, I've changed the tests to reflect that as well with `syntax: "scss"` and `syntax: "less"`.

I have also updated the title and description of the PR (including benchmarks) and the CHANGELOG to reflect these changes too.

I consider the PR ready for review again, whenever you have time.

Thanks for the help and patience.