@jeddy3 2016-07-15
The rule code looks good to me :) We'll wait on a 2nd pair of eyes to double check that bit.

> And because nesting is only possible with a preprocessor, I've changed the tests to reflect that as well with syntax: "scss" and syntax: "less".

Thanks for making the effort, but [when we say](https://github.com/stylelint/stylelint/blob/master/docs/developer-guide/rules.md#write-tests): "Ensure you use standard CSS syntax by default, and only swap parsers when testing a specific piece of non-standard syntax." we are referring to non-standard syntax like operators, `//` comments, maps and interpolation. Nesting conforms to standard CSS syntax e.g.:

```css
@media (min-width: 100px) {
  a {
    top: 10px
  }
}
```

And why there is [this proposal](https://tabatkins.github.io/specs/css-nesting/) and [corresponding PostCSS plugin](https://github.com/jonathantneal/postcss-nesting).

@ntwb's [request](https://github.com/stylelint/stylelint/pull/1659#issuecomment-232851914) was for an interpolation test e.g. `@for $n { .a { div#{$n} { } }}`.

So, just one passing `.foo { div {} }` test in the main `testRule` and one passing interpolation test in the `scss` block will suffice. You can then remove your other tests from the syntax blocks.

Don't worry, everyone's first PR goes back and forth because there's a lot of subtleties that we've evolved over the last year, especially dealing with the complexities of non-standard syntax. We've tried to document as much of it as we can, but even the documentation is subtle sometimes :)