@manovotny 2016-07-15
@jeddy3 : : 

No, this is very good. I appreciate your patience. 

I am trying to be helpful and add value to the project and trying not to suck up all of your valuable time. I appreciate the guidance and willingness to kindly nudge me in the right directions.

Yes, there are lots of subtleties, but what project doesn't have them? 😄 Part of the fun / challenge (in a good way) is figuring all that out in open source projects. 

I'll take another stab at the tests and let you know when I have them in a place where think they're ready for review again.