@davidtheclark 2016-07-16
Thanks @manovotny! I just added a couple of comments.