@manovotny 2016-07-17
Ok, @jeddy3, @ntwb, and @davidtheclark. Latest changes are ready for review.

For the tests, I believe I covered all the permutations.

* `selector-no-type` enabled, rejected
* `selector-no-type` enabled and `ignore: ["descendant"]` enabled, accepted
* `selector-no-type` enabled, scss interpolation, rejected
* `selector-no-type` enabled and `ignore: ["descendant"]` enabled, scss interpolation, accepted
* `selector-no-type` enabled, less interpolation, rejected
* `selector-no-type` enabled and `ignore: ["descendant"]` enabled, less interpolation, accepted

Please let me know if there is anything else. And please let me know if I am being dense and simply not understating what you want from a testing perspective.