@jeddy3 2016-07-17
@manovotny Thanks!

Looks thorough to me.