@manovotny 2016-07-17
Awesome! 

Thanks for all the help, @jeddy3, @ntwb, and @davidtheclark!