@manovotny 2016-07-18
Any ETA on when the next release will be cut?

No pressure. Just curious.

Thanks!