@jeddy3 2016-07-18
We were [waiting on](https://github.com/stylelint/stylelint/issues/1662) one last fix: https://github.com/stylelint/stylelint/pull/1639

It's in now and so we'll have a release out in the next few minutes :)